refactor: create bank::fee_distribution module
