+++ runtime/src/bank/tests.rs
@@ -980,232 +980,6 @@ fn test_rent_distribution() {
 }
 
 #[test]
-fn test_distribute_rent_to_validators_overflow() {
-    solana_logger::setup();
-
-    // These values are taken from the real cluster (testnet)
-    const RENT_TO_BE_DISTRIBUTED: u64 = 120_525;
-    const VALIDATOR_STAKE: u64 = 374_999_998_287_840;
-
-    let validator_pubkey = solana_sdk::pubkey::new_rand();
-    let mut genesis_config =
-        create_genesis_config_with_leader(10, &validator_pubkey, VALIDATOR_STAKE).genesis_config;
-
-    let bank = Bank::new_for_tests(&genesis_config);
-    let old_validator_lamports = bank.get_balance(&validator_pubkey);
-    bank.distribute_rent_to_validators(&bank.vote_accounts(), RENT_TO_BE_DISTRIBUTED);
-    let new_validator_lamports = bank.get_balance(&validator_pubkey);
-    assert_eq!(
-        new_validator_lamports,
-        old_validator_lamports + RENT_TO_BE_DISTRIBUTED
-    );
-
-    genesis_config
-        .accounts
-        .remove(&feature_set::no_overflow_rent_distribution::id())
-        .unwrap();
-    let bank = std::panic::AssertUnwindSafe(Bank::new_for_tests(&genesis_config));
-    let old_validator_lamports = bank.get_balance(&validator_pubkey);
-    let new_validator_lamports = std::panic::catch_unwind(|| {
-        bank.distribute_rent_to_validators(&bank.vote_accounts(), RENT_TO_BE_DISTRIBUTED);
-        bank.get_balance(&validator_pubkey)
-    });
-
-    if let Ok(new_validator_lamports) = new_validator_lamports {
-        info!("asserting overflowing incorrect rent distribution");
-        assert_ne!(
-            new_validator_lamports,
-            old_validator_lamports + RENT_TO_BE_DISTRIBUTED
-        );
-    } else {
-        info!("NOT-asserting overflowing incorrect rent distribution");
-    }
-}
-
-#[test]
-fn test_distribute_rent_to_validators_rent_paying() {
-    solana_logger::setup();
-
-    const RENT_PER_VALIDATOR: u64 = 55;
-    const TOTAL_RENT: u64 = RENT_PER_VALIDATOR * 4;
-
-    let empty_validator = ValidatorVoteKeypairs::new_rand();
-    let rent_paying_validator = ValidatorVoteKeypairs::new_rand();
-    let becomes_rent_exempt_validator = ValidatorVoteKeypairs::new_rand();
-    let rent_exempt_validator = ValidatorVoteKeypairs::new_rand();
-    let keypairs = vec![
-        &empty_validator,
-        &rent_paying_validator,
-        &becomes_rent_exempt_validator,
-        &rent_exempt_validator,
-    ];
-    let genesis_config_info = create_genesis_config_with_vote_accounts(
-        sol_to_lamports(1000.),
-        &keypairs,
-        vec![sol_to_lamports(1000.); 4],
-    );
-    let mut genesis_config = genesis_config_info.genesis_config;
-    genesis_config.rent = Rent::default(); // Ensure rent is non-zero, as genesis_utils sets Rent::free by default
-
-    for deactivate_feature in [false, true] {
-        if deactivate_feature {
-            genesis_config
-                .accounts
-                .remove(&feature_set::prevent_rent_paying_rent_recipients::id())
-                .unwrap();
-        }
-        let bank = Bank::new_for_tests(&genesis_config);
-        let rent = bank.rent_collector().rent;
-        let rent_exempt_minimum = rent.minimum_balance(0);
-
-        // Make one validator have an empty identity account
-        let mut empty_validator_account = bank
-            .get_account_with_fixed_root(&empty_validator.node_keypair.pubkey())
-            .unwrap();
-        empty_validator_account.set_lamports(0);
-        bank.store_account(
-            &empty_validator.node_keypair.pubkey(),
-            &empty_validator_account,
-        );
-
-        // Make one validator almost rent-exempt, less RENT_PER_VALIDATOR
-        let mut becomes_rent_exempt_validator_account = bank
-            .get_account_with_fixed_root(&becomes_rent_exempt_validator.node_keypair.pubkey())
-            .unwrap();
-        becomes_rent_exempt_validator_account
-            .set_lamports(rent_exempt_minimum - RENT_PER_VALIDATOR);
-        bank.store_account(
-            &becomes_rent_exempt_validator.node_keypair.pubkey(),
-            &becomes_rent_exempt_validator_account,
-        );
-
-        // Make one validator rent-exempt
-        let mut rent_exempt_validator_account = bank
-            .get_account_with_fixed_root(&rent_exempt_validator.node_keypair.pubkey())
-            .unwrap();
-        rent_exempt_validator_account.set_lamports(rent_exempt_minimum);
-        bank.store_account(
-            &rent_exempt_validator.node_keypair.pubkey(),
-            &rent_exempt_validator_account,
-        );
-
-        let get_rent_state = |bank: &Bank, address: &Pubkey| -> RentState {
-            let account = bank
-                .get_account_with_fixed_root(address)
-                .unwrap_or_default();
-            RentState::from_account(&account, &rent)
-        };
-
-        // Assert starting RentStates
-        assert_eq!(
-            get_rent_state(&bank, &empty_validator.node_keypair.pubkey()),
-            RentState::Uninitialized
-        );
-        assert_eq!(
-            get_rent_state(&bank, &rent_paying_validator.node_keypair.pubkey()),
-            RentState::RentPaying {
-                lamports: 42,
-                data_size: 0,
-            }
-        );
-        assert_eq!(
-            get_rent_state(&bank, &becomes_rent_exempt_validator.node_keypair.pubkey()),
-            RentState::RentPaying {
-                lamports: rent_exempt_minimum - RENT_PER_VALIDATOR,
-                data_size: 0,
-            }
-        );
-        assert_eq!(
-            get_rent_state(&bank, &rent_exempt_validator.node_keypair.pubkey()),
-            RentState::RentExempt
-        );
-
-        let old_empty_validator_lamports = bank.get_balance(&empty_validator.node_keypair.pubkey());
-        let old_rent_paying_validator_lamports =
-            bank.get_balance(&rent_paying_validator.node_keypair.pubkey());
-        let old_becomes_rent_exempt_validator_lamports =
-            bank.get_balance(&becomes_rent_exempt_validator.node_keypair.pubkey());
-        let old_rent_exempt_validator_lamports =
-            bank.get_balance(&rent_exempt_validator.node_keypair.pubkey());
-
-        bank.distribute_rent_to_validators(&bank.vote_accounts(), TOTAL_RENT);
-
-        let new_empty_validator_lamports = bank.get_balance(&empty_validator.node_keypair.pubkey());
-        let new_rent_paying_validator_lamports =
-            bank.get_balance(&rent_paying_validator.node_keypair.pubkey());
-        let new_becomes_rent_exempt_validator_lamports =
-            bank.get_balance(&becomes_rent_exempt_validator.node_keypair.pubkey());
-        let new_rent_exempt_validator_lamports =
-            bank.get_balance(&rent_exempt_validator.node_keypair.pubkey());
-
-        // Assert ending balances; rent should be withheld if test is active and ending RentState
-        // is RentPaying, ie. empty_validator and rent_paying_validator
-        assert_eq!(
-            if deactivate_feature {
-                old_empty_validator_lamports + RENT_PER_VALIDATOR
-            } else {
-                old_empty_validator_lamports
-            },
-            new_empty_validator_lamports
-        );
-
-        assert_eq!(
-            if deactivate_feature {
-                old_rent_paying_validator_lamports + RENT_PER_VALIDATOR
-            } else {
-                old_rent_paying_validator_lamports
-            },
-            new_rent_paying_validator_lamports
-        );
-
-        assert_eq!(
-            old_becomes_rent_exempt_validator_lamports + RENT_PER_VALIDATOR,
-            new_becomes_rent_exempt_validator_lamports
-        );
-
-        assert_eq!(
-            old_rent_exempt_validator_lamports + RENT_PER_VALIDATOR,
-            new_rent_exempt_validator_lamports
-        );
-
-        // Assert ending RentStates
-        assert_eq!(
-            if deactivate_feature {
-                RentState::RentPaying {
-                    lamports: RENT_PER_VALIDATOR,
-                    data_size: 0,
-                }
-            } else {
-                RentState::Uninitialized
-            },
-            get_rent_state(&bank, &empty_validator.node_keypair.pubkey()),
-        );
-        assert_eq!(
-            if deactivate_feature {
-                RentState::RentPaying {
-                    lamports: old_rent_paying_validator_lamports + RENT_PER_VALIDATOR,
-                    data_size: 0,
-                }
-            } else {
-                RentState::RentPaying {
-                    lamports: old_rent_paying_validator_lamports,
-                    data_size: 0,
-                }
-            },
-            get_rent_state(&bank, &rent_paying_validator.node_keypair.pubkey()),
-        );
-        assert_eq!(
-            RentState::RentExempt,
-            get_rent_state(&bank, &becomes_rent_exempt_validator.node_keypair.pubkey()),
-        );
-        assert_eq!(
-            RentState::RentExempt,
-            get_rent_state(&bank, &rent_exempt_validator.node_keypair.pubkey()),
-        );
-    }
-}
-
-#[test]
 fn test_rent_exempt_executable_account() {
     let (mut genesis_config, mint_keypair) = create_genesis_config(100_000);
     genesis_config.rent = rent_with_exemption_threshold(1000.0);
